refactor(settings): move service helpers to dedicated package
