+++ core/metest/helpers_services.go
@@ -1,4 +1,4 @@
-package services_test
+package metest
 
 import (
 	"context"
